requirements.txt
Django==5.2.7
djangorestframework==3.16.1
drf-spectacular==0.28.0
mysqlclient==2.2.7
pillow==12.0.0
python-dotenv==1.1.1
requests==2.32.5